+++ src/tests.zig
@@ -5,4 +5,5 @@ comptime {
     _ = @import("chess/BitBoard.zig");
     _ = @import("chess/bit_board/LookupImpl.zig");
     _ = @import("chess/bit_board/MathImpl.zig");
+    _ = @import("chess/color.zig");
 }